# fate-manager/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "fate manager api documents",
        "title": "Fate-Manager",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:9090",
    "paths": {
...
        "/fate-manager/api/service/action": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ServiceController"
                ],
                "summary": "Do stopped Or Running Action",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.ActionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.CommResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/fate-manager/api/service/connectkubefate": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ServiceController"
                ],
                "summary": "Connect Exists Kubernetes",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.KubeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.CommResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/fate-manager/api/service/info": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ServiceController"
                ],
                "summary": "Get Service Info",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.ServiceInfoReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.ServiceInfoResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/fate-manager/api/service/installstatus": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ServiceController"
                ],
                "summary": "Get PartyId If Install",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.InstallStatus"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/fate-manager/api/service/log": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ServiceController"
                ],
                "summary": "Component Install Log",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.LogReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.LogResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/fate-manager/api/service/overview": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ServiceController"
                ],
                "summary": "Get Service Overview List",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.OverViewReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.OverViewResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/fate-manager/api/service/pagestatus": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ServiceController"
                ],
                "summary": "Get Page Status",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.PageStatusReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.GetPageStatusResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/fate-manager/api/service/upgradelist": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ServiceController"
                ],
                "summary": "Get Upgrade Fate Version List",
                "parameters": [
                    {
                        "description": "request param",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.UpgradeFateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.UpgradeFateResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "app.CommResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "app.GetPageStatusResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/entity.PageStatusResp"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "app.LogResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "app.OverViewResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.OverViewRspItem"
                    }
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "app.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "app.ServiceInfoResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.ComponentDeploy"
                    }
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "app.UpgradeFateResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.UpdateVersionResp"
                    }
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "entity.ActionReq": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "integer"
                },
                "componentName": {
                    "type": "string"
                },
                "federatedId": {
                    "type": "integer"
                },
                "partyId": {
                    "type": "integer"
                },
                "productType": {
                    "type": "integer"
                }
            }
        },
...
        "entity.ComponentDeploy": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "componentName": {
                    "type": "string"
                },
                "componentVersion": {
                    "type": "string"
                },
                "deployStatus": {
                    "type": "object",
                    "$ref": "#/definitions/entity.IdPair"
                },
                "duration": {
                    "type": "integer"
                },
                "finishTime": {
                    "type": "integer"
                },
                "siteName": {
                    "type": "string"
                },
                "startTime": {
                    "type": "integer"
                },
                "status": {
                    "type": "object",
                    "$ref": "#/definitions/entity.IdPair"
                }
            }
        },
...
        "entity.IdPair": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "desc": {
                    "type": "string"
                }
            }
        },
...
        "entity.InstallItem": {
            "type": "object",
            "properties": {
                "ComponentVersion": {
                    "type": "string"
                },
                "componentName": {
                    "type": "string"
                },
                "installTime": {
                    "type": "integer"
                },
                "operation": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Operation"
                    }
                },
                "upgradeStatus": {
                    "type": "boolean"
                },
                "upgradeTime": {
                    "type": "integer"
                }
            }
        },
...
        "entity.InstallStatus": {
            "type": "object",
            "properties": {
                "federatedId": {
                    "type": "integer"
                },
                "partyId": {
                    "type": "integer"
                },
                "productType": {
                    "type": "integer"
                }
            }
        },
        "entity.KubeReq": {
            "type": "object",
            "properties": {
                "federatedId": {
                    "type": "integer"
                },
                "kubenetesUrl": {
                    "type": "string"
                },
                "partyId": {
                    "type": "integer"
                },
                "productType": {
                    "type": "integer"
                }
            }
        },
        "entity.LogReq": {
            "type": "object",
            "properties": {
                "federatedId": {
                    "type": "integer"
                },
                "partyId": {
                    "type": "integer"
                },
                "productType": {
                    "type": "integer"
                }
            }
        },
...
        "entity.Operation": {
            "type": "object",
            "properties": {
                "operateAction": {
                    "type": "object",
                    "$ref": "#/definitions/entity.IdPair"
                },
                "operateTime": {
                    "type": "integer"
                }
            }
        },
        "entity.OverViewReq": {
            "type": "object",
            "properties": {
                "condition": {
                    "type": "string"
                },
                "federatedId": {
                    "type": "integer"
                },
                "role": {
                    "type": "integer"
                }
            }
        },
        "entity.OverViewRspItem": {
            "type": "object",
            "properties": {
                "fateVersion": {
                    "type": "string"
                },
                "federatedId": {
                    "type": "integer"
                },
                "federatedOrganization": {
                    "type": "string"
                },
                "installInfo": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.InstallItem"
                    }
                },
                "institutions": {
                    "type": "string"
                },
                "partyId": {
                    "type": "integer"
                },
                "role": {
                    "type": "object",
                    "$ref": "#/definitions/entity.IdPair"
                },
                "siteName": {
                    "type": "string"
                }
            }
        },
        "entity.PageStatusReq": {
            "type": "object",
            "properties": {
                "federatedId": {
                    "type": "integer"
                },
                "partyId": {
                    "type": "integer"
                },
                "productType": {
                    "type": "integer"
                }
            }
        },
        "entity.PageStatusResp": {
            "type": "object",
            "properties": {
                "pageStatus": {
                    "type": "object",
                    "$ref": "#/definitions/entity.IdPair"
                }
            }
        },
...
        "entity.ServiceInfoReq": {
            "type": "object",
            "properties": {
                "fateVersion": {
                    "type": "string"
                },
                "federatedId": {
                    "type": "integer"
                },
                "partyId": {
                    "type": "integer"
                },
                "productType": {
                    "type": "integer"
                }
            }
        },
...
        "entity.UpdateVersionResp": {
            "type": "object",
            "properties": {
                "fateVersion": {
                    "type": "string"
                },
                "versionIndex": {
                    "type": "integer"
                }
            }
        },
        "entity.UpgradeFateReq": {
            "type": "object",
            "properties": {
                "federatedId": {
                    "type": "integer"
                },
                "partyId": {
                    "type": "integer"
                },
                "productType": {
                    "type": "integer"
                }
            }
        },
...
    }
}